Handle uploading S3 files in the service
