+++ publishing-example-app/src/main/java/com/ably/tracking/example/publisher/PublisherServiceActivity.kt
@@ -1,33 +1,19 @@
 package com.ably.tracking.example.publisher
 
-import android.Manifest
 import android.content.ComponentName
 import android.content.Intent
 import android.content.ServiceConnection
 import android.os.Bundle
 import android.os.IBinder
-import androidx.annotation.RequiresPermission
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
-import com.ably.tracking.Accuracy
-import com.ably.tracking.ConnectionConfiguration
-import com.ably.tracking.Resolution
-import com.ably.tracking.publisher.DefaultResolutionPolicyFactory
-import com.ably.tracking.publisher.LocationSource
-import com.ably.tracking.publisher.MapConfiguration
-import com.ably.tracking.publisher.Publisher
-import com.ably.tracking.publisher.RoutingProfile
 
 private const val NO_FLAGS = 0
-private const val MAPBOX_ACCESS_TOKEN = BuildConfig.MAPBOX_ACCESS_TOKEN
-private const val CLIENT_ID = "<INSERT_CLIENT_ID_HERE>"
-private const val ABLY_API_KEY = BuildConfig.ABLY_API_KEY
 
 /**
  * The activity to extend when connection to the publisher service is needed.
  */
 abstract class PublisherServiceActivity : AppCompatActivity() {
-    private val DEFAULT_RESOLUTION = Resolution(Accuracy.MINIMUM, 1000L, 1.0)
     protected var publisherService: PublisherService? = null
     private val publisherServiceConnection = object : ServiceConnection {
         override fun onServiceConnected(className: ComponentName, serviceBinder: IBinder) {
@@ -51,24 +37,6 @@ abstract class PublisherServiceActivity : AppCompatActivity() {
     override fun onDestroy() {
         unbindPublisherService()
         super.onDestroy()
-    }
-
-    /**
-     * Creates and starts the [Publisher] from the [PublisherService].
-     */
-    @RequiresPermission(anyOf = [Manifest.permission.ACCESS_COARSE_LOCATION, Manifest.permission.ACCESS_FINE_LOCATION])
-    protected fun startPublisher(
-        defaultResolution: Resolution = DEFAULT_RESOLUTION,
-        locationSource: LocationSource? = null
-    ) {
-        publisherService?.publisher = Publisher.publishers()
-            .connection(ConnectionConfiguration(ABLY_API_KEY, CLIENT_ID))
-            .map(MapConfiguration(MAPBOX_ACCESS_TOKEN))
-            .locationSource(locationSource)
-            .resolutionPolicy(DefaultResolutionPolicyFactory(defaultResolution, this))
-            .androidContext(this)
-            .profile(RoutingProfile.DRIVING)
-            .start()
     }
 
     /**
